+++ src/app/(layout-with-navigation)/(simulation)/fin-v2/_components/results/TotalCard.tsx
