+++ src/app/(layout-with-navigation)/(simulation)/fin/_components/Slide.tsx
