+++ src/app/(layout-with-navigation)/(simulation)/fin/_components/Subcategories.tsx
