+++ src/app/(layout-with-navigation)/(simulation)/fin/_components/gridChart/ShareImage.tsx
