+++ src/app/(layout-with-navigation)/(simulation)/fin/_components/totalVsTarget/Total.tsx
