+++ src/app/(layout-with-navigation)/(simulation)/fin/page.tsx
@@ -1,14 +1,12 @@
 import IframeDataShareModal from '@/components/iframe/IframeDataShareModal'
-import Trans from '@/components/translation/Trans'
-import Separator from '@/design-system/layout/Separator'
+import NorthStarBanner from '@/components/northstar/NorthstarBanner'
+import ButtonLink from '@/design-system/inputs/ButtonLink'
 import { getMetadataObject } from '@/helpers/metadata/getMetadataObject'
 import { FormProvider } from '@/publicodes-state'
-import CongratulationsText from './_components/CongratulationsText'
-import FeedbackBanner from './_components/FeedbackBanner'
-import GetResultsByEmail from './_components/GetResultsByEmail'
-import GroupModePromotionBanner from './_components/GroupModePromotionBanner'
-import HowToAct from './_components/HowToAct'
-import Results from './_components/Results'
+import { Diapo } from '@/types/fin'
+import FinSlider from './_components/FinSlider'
+import { NewsletterForm } from './_components/NewsletterForm'
+import './slick.css'
 
 export async function generateMetadata() {
   return getMetadataObject({
@@ -18,33 +16,26 @@ export async function generateMetadata() {
   })
 }
 
-export default function FinPage() {
+export default function FinPage({
+  searchParams,
+}: {
+  searchParams: Record<string, Diapo>
+}) {
   return (
     <FormProvider>
+      <NorthStarBanner type="learned" />
+
       <IframeDataShareModal />
 
-      <CongratulationsText />
+      <div className="mb-12 flex justify-start md:mx-16">
+        <ButtonLink size="sm" color="secondary" href="/simulateur/bilan">
+          ← Revenir au test
+        </ButtonLink>
+      </div>
 
-      <Results />
+      <FinSlider searchParams={searchParams} />
 
-      <GetResultsByEmail />
-
-      <GroupModePromotionBanner />
-
-      <Separator className="my-8" />
-
-      <HowToAct />
-
-      <FeedbackBanner
-        className="mb-8 mt-12"
-        text={
-          <Trans i18nKey="publicodes.northstar.learned">
-            Est-ce que "Nos Gestes Climat" vous a permis d'apprendre quelque
-            chose ?
-          </Trans>
-        }
-        type="learned"
-      />
+      <NewsletterForm />
     </FormProvider>
   )
 }
