+++ src/app/(layout-with-navigation)/(simulation)/fin/slick.css
