+++ src/constants/ab-tests.ts
@@ -1,0 +1,8 @@
+export const AB_TESTS = [
+  {
+    name: 'fin',
+    stringToMatch: '/fin',
+    newPath: '/fin-v2',
+    oldPath: '/fin',
+  },
+]
